requirements.txt
Django==1.6.4
FeinCMS==1.9.3
MySQL-python==1.2.5
Pillow==2.4.0
Unidecode==0.04.14
dj-database-url==0.3.0
dj-static==0.0.5
django-debug-toolbar==1.2
django-mptt==0.6.0
django-pagination==1.0.7
django-suit==0.2.8
django-suit-redactor==0.0.2
django-tastypie==0.11.0
django-uuslug==1.0.2
easy-thumbnails==2.0.1
feedparser==5.1.3
gunicorn==19.0.0
pystache==0.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
python-slugify==0.0.7
six==1.6.1
sqlparse==0.1.11
static==1.0.2
virtualenv==1.11.6
wsgiref==0.1.2
